Hunting Record System packet implementation

+++ WebZen.DataBase/CharacterDto.cs
@@ -24,7 +24,7 @@ namespace MU.DataBase
         public GuildDto Guild { get; set; }
 
         [Column(TypeName = "SMALLINT(5)")]
-        public byte Map { get; set; }
+        public ushort Map { get; set; }
         public short X { get; set; }
         public short Y { get; set; }
